Make mark_aromatic_atoms more aggressive
It now de-kekulizes all regions that show DIME, and kekulizes all regions that don't.
Part of this change is the removal of the `atoms` argument to the function, and introduction of the `strict` argument.
The removal of the `atoms` argument also changed the behaviour of 2 test cases, where all non-saturated atoms would be eligible for making aromatic regions.